--- lib/auth/data/auth/login_response/login_response.dart
@@ -1,17 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:moprog/auth/data/auth/refresh_token_response/refresh_token_response.dart';
-import 'package:moprog/auth/data/auth/user/user.dart';
-
-part 'login_response.freezed.dart';
-part 'login_response.g.dart';
-
-@freezed
-abstract class LoginResponse with _$LoginResponse {
-  factory LoginResponse({
-    required RefreshTokenResponse tokens,
-    required User user,
-  }) = _LoginResponse;
-
-  factory LoginResponse.fromJson(Map<String, dynamic> json) =>
-      _$LoginResponseFromJson(json);
-}
