+++ lib/auth/domain/auth_method.dart
