+++ lib/auth/domain/check_auth_status.dart
@@ -1,0 +1,7 @@
+enum CheckAuthStatus {
+  authenticated,
+  unauthenticated,
+  noInternet,
+  serverError,
+  needVerification
+}
