+++ lib/core/data/utils/launched_effect.dart
