--- lib/core/utils/view_model.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/foundation.dart';
-
-typedef ViewModel = ChangeNotifier;
