+++ lib/router.dart
@@ -1,9 +1,9 @@
 import 'package:go_router/go_router.dart';
-import 'package:moprog/auth/presentation/sign_in/sign_in_screen.dart';
-import 'package:moprog/core/utils/navigator.dart';
-import 'package:moprog/auth/presentation/sign_up/sign_up_screen.dart';
-import 'package:moprog/core/presentation/splash/splash_screen.dart';
-import 'package:moprog/main/presentation/home/home_screen.dart';
+import 'package:moprog/auth/presentation/pages/sign_in_screen.dart';
+import 'package:moprog/auth/presentation/pages/sign_up_screen.dart';
+import 'package:moprog/core/data/utils/navigator.dart';
+import 'package:moprog/auth/presentation/pages/splash_screen.dart';
+import 'package:moprog/main/home/presentation/pages/home_screen.dart';
 
 /// Navigasi dari mana ke mana langsung diatur dari sini, biar terpusat dan gampang debug nya
 final router = GoRouter(
